requirements.txt
fastapi>=0.110
uvicorn>=0.29
pydantic>=2.6
pytest>=8.0
httpx>=0.26
SQLAlchemy>=2.0
python-jose>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.9
psycopg>=3.1
python-dotenv>=1.0
email-validator>=2.1
